Relocates package_test.go from state_test into the state package. Implements export_test.go in package_test.go and deletes it.
